create PollMixin, relocate PollManager, relocate Closed
